+++ src/components/FormInputs/index.js
@@ -1,3 +1,2 @@
 export { default as Input } from './Input';
-export { default as CardPasswordInput } from './CardPasswordInput';
 export { default as InputWithCounter } from './InputWithCounter';
