+++ src/components/index.js
@@ -1,10 +1,7 @@
-export { default as CardCompanySelection } from './CardCompanySelection';
-export { default as SecurityCodeGuide } from './SecurityCodeGuide';
 export { TextButton } from './Buttons';
 export { default as Card } from './Card';
 export { default as AddCard } from './AddCard';
-export { default as CardCompanyOption } from './CardCompanyOption';
-export { Input, CardPasswordInput, InputWithCounter } from './FormInputs';
+export { Input, InputWithCounter } from './FormInputs';
 export { default as Icon } from './Icon';
 export { default as Modal } from './Modal';
 export { default as Header } from './Header';
