+++ src/pages/AddCard/Form/SecurityCode/SecurityCodeGuide/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import securityGuideImage from '../../assets/securityGuideImage.svg';
+import securityGuideImage from '../../../../../assets/securityGuideImage.svg';
 
 export default function SecurityCodeGuide() {
   return <img src={securityGuideImage} alt="securityGuideImage" width="280px"></img>;
