+++ src/pages/index.js
@@ -1,2 +1,2 @@
-export { default as AddCardForm } from './AddCardForm';
+export { default as AddCard } from './AddCard';
 export { default as AddCardComplete } from './AddCardComplete';
